backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
langchain==0.3.0
langchain-openai==0.2.0
langgraph==0.2.40
chromadb==0.5.0
sentence-transformers==2.7.0
pymupdf==1.24.0
python-dotenv==1.0.1
aiofiles==24.1.0
transformers==4.40.0
torch==2.2.0
accelerate==0.30.0
pytesseract==0.3.13
Pillow==10.4.0
easyocr==1.7.1

